Pulled out refactor changes.
