+++ src/ADP.Portal.Core/Git/Infrastructure/IGitOpsConfigRepository.cs
@@ -13,16 +13,14 @@ namespace ADP.Portal.Core.Git.Infrastructure
 
         Task<IEnumerable<KeyValuePair<string, Dictionary<object, object>>>> GetAllFilesAsync(GitRepo gitRepo, string path);
 
-        Task PushFilesToRepository(GitRepo gitRepoFluxServices, string branchName, Dictionary<string, Dictionary<object, object>> generatedFiles);
+        Task<Reference?> GetBranchAsync(GitRepo gitRepo, string branchName);
 
-        //Task<Reference?> GetBranchAsync(GitRepo gitRepo, string branchName);
+        Task<Reference> CreateBranchAsync(GitRepo gitRepo, string branchName, string sha);
 
-        //Task<Reference> CreateBranchAsync(GitRepo gitRepo, string branchName, string sha);
+        Task<Reference> UpdateBranchAsync(GitRepo gitRepo, string branchName, string sha);
 
-        //Task<Reference> UpdateBranchAsync(GitRepo gitRepo, string branchName, string sha);
+        Task<Commit?> CreateCommitAsync(GitRepo gitRepo, Dictionary<string, Dictionary<object, object>> generatedFiles, string message, string? branchName = null);
 
-        //Task<Commit?> CreateCommitAsync(GitRepo gitRepo, Dictionary<string, Dictionary<object, object>> generatedFiles, string message, string? branchName = null);
-
-        //Task<bool> CreatePullRequestAsync(GitRepo gitRepo, string branchName, string message);
+        Task<bool> CreatePullRequestAsync(GitRepo gitRepo, string branchName, string message);
     }
 }
